cli/requirements.txt
Arpeggio==1.10.2
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
dill==0.3.4
elasticsearch==7.13.4
ftfy==6.0.3
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.1
grpcio==1.39.0
idna==3.2
iniconfig==1.1.1
kubernetes==17.17.0
mailchecker==4.0.11
oauthlib==3.1.1
packaging==21.0
parver==0.3.1
phonenumbers==8.12.28
pluggy==0.13.1
protobuf==3.17.3
pulumi==3.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.4
python-benedict==0.24.0
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
semver==2.13.0
six==1.16.0
smmap==4.0.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==1.1.0
xmltodict==0.12.0
